[Ref] Refactor list-related integration tests into separate class.
