+++ jopa-integration-tests/src/main/java/cz/cvut/kbss/jopa/test/runner/CreateOperationsRunner.java
@@ -21,11 +21,30 @@ import cz.cvut.kbss.jopa.exception.IdentifierNotSetException;
 import cz.cvut.kbss.jopa.exceptions.OWLEntityExistsException;
 import cz.cvut.kbss.jopa.exceptions.RollbackException;
 import cz.cvut.kbss.jopa.model.EntityManager;
-import cz.cvut.kbss.jopa.model.SequencesVocabulary;
 import cz.cvut.kbss.jopa.model.descriptors.Descriptor;
 import cz.cvut.kbss.jopa.model.descriptors.EntityDescriptor;
-import cz.cvut.kbss.jopa.test.*;
-import cz.cvut.kbss.jopa.test.environment.*;
+import cz.cvut.kbss.jopa.test.OWLClassA;
+import cz.cvut.kbss.jopa.test.OWLClassB;
+import cz.cvut.kbss.jopa.test.OWLClassD;
+import cz.cvut.kbss.jopa.test.OWLClassE;
+import cz.cvut.kbss.jopa.test.OWLClassG;
+import cz.cvut.kbss.jopa.test.OWLClassH;
+import cz.cvut.kbss.jopa.test.OWLClassK;
+import cz.cvut.kbss.jopa.test.OWLClassM;
+import cz.cvut.kbss.jopa.test.OWLClassN;
+import cz.cvut.kbss.jopa.test.OWLClassP;
+import cz.cvut.kbss.jopa.test.OWLClassWithQueryAttr;
+import cz.cvut.kbss.jopa.test.OWLClassWithQueryAttr2;
+import cz.cvut.kbss.jopa.test.OWLClassWithQueryAttr3;
+import cz.cvut.kbss.jopa.test.OWLClassWithQueryAttr4;
+import cz.cvut.kbss.jopa.test.OWLClassWithQueryAttr5;
+import cz.cvut.kbss.jopa.test.OWLClassX;
+import cz.cvut.kbss.jopa.test.Vocabulary;
+import cz.cvut.kbss.jopa.test.environment.DataAccessor;
+import cz.cvut.kbss.jopa.test.environment.Generators;
+import cz.cvut.kbss.jopa.test.environment.PersistenceFactory;
+import cz.cvut.kbss.jopa.test.environment.Quad;
+import cz.cvut.kbss.jopa.test.environment.TestEnvironment;
 import cz.cvut.kbss.jopa.vocabulary.XSD;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
@@ -36,11 +55,23 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.ZoneOffset;
 import java.time.temporal.ChronoUnit;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public abstract class CreateOperationsRunner extends BaseRunner {
 
@@ -123,128 +154,6 @@ public abstract class CreateOperationsRunner extends BaseRunner {
         em.getTransaction().begin();
         em.detach(det);
         assertThrows(OWLEntityExistsException.class, () -> em.persist(det));
-    }
-
-    @Test
-    void testPersistWithSimpleList() {
-        this.em = getEntityManager("PersistSimpleList", false);
-        entityC.setSimpleList(Generators.createSimpleList(10));
-        persistCWithLists(entityC);
-
-        final OWLClassA a = findRequired(OWLClassA.class, entityC.getSimpleList().get(1).getUri());
-        final OWLClassC c = findRequired(OWLClassC.class, entityC.getUri());
-        assertNotNull(c.getSimpleList());
-        assertFalse(c.getSimpleList().isEmpty());
-        assertEquals(entityC.getSimpleList().size(), c.getSimpleList().size());
-        assertTrue(c.getSimpleList().contains(a));
-    }
-
-    @Test
-    void persistingEntityWithSimpleListWithoutCascadeIsIllegal() {
-        this.em = getEntityManager("PersistSimpleListNoCascade", false);
-        entityC.setSimpleList(Generators.createSimpleList(10));
-        assertThrows(RollbackException.class, () -> persist(entityC));
-    }
-
-    @Test
-    void persistWithSimpleListSavesListReferenceWhenAllItemsArePersisted() {
-        this.em = getEntityManager("persistWithSimpleListSavesListReferenceWhenAllItemsArePersisted", false);
-        final OWLClassK entityK = new OWLClassK();
-        entityK.setSimpleList(IntStream.range(0, 5).mapToObj(i -> {
-            final OWLClassE item = new OWLClassE();
-            item.setStringAttribute("item" + i);
-            return item;
-        }).collect(Collectors.toList()));
-        em.getTransaction().begin();
-        em.persist(entityK);
-        entityK.getSimpleList().forEach(item -> assertNull(item.getUri()));
-        entityK.getSimpleList().forEach(em::persist);
-        entityK.getSimpleList().forEach(item -> assertNotNull(item.getUri()));
-        em.getTransaction().commit();
-
-        final OWLClassK result = findRequired(OWLClassK.class, entityK.getUri());
-        verifyLists(entityK.getSimpleList(), result.getSimpleList());
-    }
-
-    private static void verifyLists(List<OWLClassE> expected, List<OWLClassE> actual) {
-        assertEquals(expected.size(), actual.size());
-        for (int i = 0; i < expected.size(); i++) {
-            assertEquals(expected.get(i).getUri(), actual.get(i).getUri());
-            assertEquals(expected.get(i).getStringAttribute(), actual.get(i).getStringAttribute());
-        }
-    }
-
-    @Test
-    void testPersistWithReferencedList() {
-        this.em = getEntityManager("PersistReferencedList", false);
-        entityC.setReferencedList(Generators.createReferencedList(5));
-        em.getTransaction().begin();
-        em.persist(entityC);
-        entityC.getReferencedList().forEach(em::persist);
-        assertTrue(em.contains(entityC));
-        assertTrue(em.contains(entityC.getReferencedList().get(0)));
-        em.getTransaction().commit();
-
-        final OWLClassC c = findRequired(OWLClassC.class, entityC.getUri());
-        assertNotNull(c.getReferencedList());
-        assertFalse(c.getReferencedList().isEmpty());
-        assertEquals(entityC.getReferencedList().size(), c.getReferencedList().size());
-        for (OWLClassA a : entityC.getReferencedList()) {
-            final OWLClassA resA = em.find(OWLClassA.class, a.getUri());
-            assertNotNull(resA);
-            assertEquals(a.getStringAttribute(), resA.getStringAttribute());
-            assertTrue(c.getReferencedList().contains(resA));
-        }
-    }
-
-
-    @Test
-    void persistingEntityWithReferencedListWithoutCascadeIsIllegal() {
-        this.em = getEntityManager("PersistReferencedListNoCascade", false);
-        entityC.setReferencedList(Generators.createReferencedList(5));
-        assertThrows(RollbackException.class, () -> persist(entityC));
-    }
-
-    @Test
-    void persistWithReferencedListSavesListReferenceWhenAllItemsArePersisted() {
-        this.em = getEntityManager("persistWithReferencedListSavesListReferenceWhenAllItemsArePersisted", false);
-        final OWLClassK entityK = new OWLClassK();
-        entityK.setReferencedList(IntStream.range(0, 5).mapToObj(i -> {
-            final OWLClassE item = new OWLClassE();
-            item.setStringAttribute("item" + i);
-            return item;
-        }).collect(Collectors.toList()));
-        em.getTransaction().begin();
-        em.persist(entityK);
-        entityK.getReferencedList().forEach(item -> assertNull(item.getUri()));
-        entityK.getReferencedList().forEach(em::persist);
-        entityK.getReferencedList().forEach(item -> assertNotNull(item.getUri()));
-        em.getTransaction().commit();
-
-        final OWLClassK result = findRequired(OWLClassK.class, entityK.getUri());
-        verifyLists(entityK.getReferencedList(), result.getReferencedList());
-    }
-
-    @Test
-    void testPersistSimpleAndReferencedList() {
-        this.em = getEntityManager("PersistSimpleAndReferencedList", false);
-        entityC.setReferencedList(Generators.createReferencedList(5));
-        entityC.setSimpleList(Generators.createSimpleList(5));
-        persistCWithLists(entityC);
-
-        final OWLClassC c = findRequired(OWLClassC.class, entityC.getUri());
-        assertNotNull(c.getSimpleList());
-        assertEquals(entityC.getSimpleList().size(), c.getSimpleList().size());
-        assertNotNull(c.getReferencedList());
-        assertEquals(entityC.getReferencedList().size(), c.getReferencedList().size());
-        for (OWLClassA a : entityC.getSimpleList()) {
-            final OWLClassA resA = findRequired(OWLClassA.class, a.getUri());
-            assertTrue(c.getSimpleList().contains(resA));
-        }
-        for (OWLClassA a : entityC.getReferencedList()) {
-            final OWLClassA resA = findRequired(OWLClassA.class, a.getUri());
-            assertTrue(c.getReferencedList().contains(resA));
-        }
     }
 
     @Test
@@ -397,28 +306,6 @@ public abstract class CreateOperationsRunner extends BaseRunner {
 
         final OWLClassP res = findRequired(OWLClassP.class, entityP.getUri());
         assertEquals(urls, res.getIndividuals());
-    }
-
-    @Test
-    void testPersistInstanceWithSimpleListOfIdentifiers() {
-        this.em = getEntityManager("PersistInstanceWithSimpleListOfIdentifiers", false);
-        entityP.setSimpleList(Generators.createListOfIdentifiers());
-        persist(entityP);
-        em.clear();
-
-        final OWLClassP res = findRequired(OWLClassP.class, entityP.getUri());
-        assertEquals(entityP.getSimpleList(), res.getSimpleList());
-    }
-
-    @Test
-    void testPersistInstanceWithReferencedListOfIdentifiers() {
-        this.em = getEntityManager("PersistInstanceWithReferencedListOfIdentifiers", false);
-        entityP.setReferencedList(Generators.createListOfIdentifiers());
-        persist(entityP);
-        em.clear();
-
-        final OWLClassP res = findRequired(OWLClassP.class, entityP.getUri());
-        assertEquals(entityP.getReferencedList(), res.getReferencedList());
     }
 
     @Test
@@ -790,19 +677,5 @@ public abstract class CreateOperationsRunner extends BaseRunner {
                         entityM.getAnnotationSimpleLiteral(), (String) null),
                 new Quad(URI.create(entityM.getKey()), URI.create(Vocabulary.p_m_simpleLiteral),
                         entityM.getSimpleLiteral(), (String) null)), em);
-    }
-
-    @Test
-    void persistSupportsReferencedListsContainingDataPropertyLiteralValues() {
-        this.em = getEntityManager("persistSupportsReferencedListsContainingDataPropertyLiteralValues", false);
-        entityM.setLiteralReferencedList(Generators.createDataPropertyList());
-        persist(entityM);
-
-        for (int i = 0; i < entityM.getLiteralReferencedList().size(); i++) {
-            assertTrue(em.createNativeQuery("ASK WHERE { ?prev ?hasNext ?node . ?node ?hasContent ?content . }", Boolean.class)
-                         .setParameter("hasNext", URI.create(i == 0 ? Vocabulary.p_m_literalReferencedList : SequencesVocabulary.s_p_hasNext))
-                         .setParameter("hasContent", URI.create(SequencesVocabulary.s_p_hasContents))
-                         .setParameter("content", entityM.getLiteralReferencedList().get(i)).getSingleResult());
-        }
     }
 }
